+++ gradle.properties
@@ -12,7 +12,7 @@ org.gradle.jvmargs=-Xmx2G
 	fabric_version=0.58.6+1.19.2
 
 # Mod Properties
-	mod_version=1.9.3
+	mod_version=1.10.0
 	# todo change
 	maven_group = azzy.fabric
 	archives_base_name = incubus-core
