+++ src/main/java/net/id/incubus_core/recipe/RecipeParser.java
@@ -1,32 +1,21 @@
-package net.id.incubus_core.json;
+package net.id.incubus_core.recipe;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.MalformedJsonException;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import net.id.incubus_core.IncubusCore;
-import net.id.incubus_core.recipe.IngredientStack;
-import net.id.incubus_core.recipe.OptionalStack;
-import net.id.incubus_core.recipe.matchbook.MatchRegistry;
-import net.id.incubus_core.recipe.matchbook.Matchbook;
-import net.id.incubus_core.util.RegistryHelper;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.nbt.NbtHelper;
-import net.minecraft.recipe.Ingredient;
-import net.minecraft.tag.TagKey;
-import net.minecraft.util.Identifier;
-import net.minecraft.util.registry.Registry;
+import com.google.gson.*;
+import com.google.gson.stream.*;
+import com.mojang.brigadier.exceptions.*;
+import net.id.incubus_core.*;
+import net.id.incubus_core.recipe.matchbook.*;
+import net.id.incubus_core.util.*;
+import net.minecraft.item.*;
+import net.minecraft.nbt.*;
+import net.minecraft.recipe.*;
+import net.minecraft.tag.*;
+import net.minecraft.util.*;
+import net.minecraft.util.registry.*;
 
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
+import java.io.*;
+import java.nio.charset.*;
+import java.util.*;
 
 @SuppressWarnings("unused")
 public class RecipeParser {
@@ -140,6 +129,38 @@ public class RecipeParser {
         }
 
         return builder.build(mode);
+    }
+
+    /**
+     * Parses an ItemStack json object with optional NBT data
+     * Can be used in RecipeSerializers to get ItemStacks with NBT as output
+     * @param json The JsonObject to parse
+     * @return An ItemStack with nbt data, like specified in the json
+     */
+    public static ItemStack getItemStackWithNbtFromJson(JsonObject json) {
+        Item item = ShapedRecipe.getItem(json);
+        if (json.has("data")) {
+            throw new JsonParseException("Disallowed data tag found");
+        } else {
+            int count = JsonHelper.getInt(json, "count", 1);
+            if (count < 1) {
+                throw new JsonSyntaxException("Invalid output count: " + count);
+            } else {
+                ItemStack stack = new ItemStack(item, count);
+                String nbt = JsonHelper.getString(json, "nbt", "");
+                if (!nbt.isEmpty()) {
+                    try {
+                        NbtCompound compound = NbtHelper.fromNbtProviderString(nbt);
+                        compound.remove("palette");
+                        stack.setNbt(compound);
+                    } catch (CommandSyntaxException e) {
+                        throw new JsonSyntaxException("Invalid output nbt: " + nbt);
+                    }
+                }
+
+                return stack;
+            }
+        }
     }
 
 }
